BACKEND/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
filelock==3.13.1
fsspec==2024.2.0
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
sympy==1.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5